+++ test/resampling/resampling_schemes/test_RandomSequences.jl
@@ -1,0 +1,2 @@
+r = RandomSequences(10, 10)
+@test r isa RandomSequences